Brain-Tumor-Detection-App/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
numpy==1.24.3
pillow==9.5.0
werkzeug==2.0.3
matplotlib==3.7.1
scikit-learn==1.2.2
opencv-python==4.7.0.72